save err check

+++ app/src/main/java/com/dopy/dopy/tayga/model/twich/TwitchStream.java
@@ -18,7 +18,7 @@ public class TwitchStream extends BroadcastModel {
 
     @SerializedName("_id")
     @Expose
-    public Integer id;
+    public double id;
     @SerializedName("game")
     @Expose
     public String game;
